aplicacion/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
azure-core==1.30.2
azure-data-tables==12.5.0
azure-functions==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
frozenlist==1.4.1
idna==3.7
install==1.3.5
ipykernel==6.29.4
ipython==8.24.0
isodate==0.6.1
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.0.5
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pyzmq==26.0.3
requests==2.31.0
sendgrid==6.11.0
six==1.16.0
stack-data==0.6.3
starkbank-ecdsa==2.2.0
tornado==6.4
traitlets==5.14.3
twilio==9.0.5
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
